awfulordersystemBackEnd/requirements.txt
pywin32~=303
flask~=2.0.3
numpy~=1.21.5
pandas~=1.4.2
pillow~=9.0.1